+++ index.ts
@@ -1,5 +1,6 @@
 export { YINModule } from './src/yin/yin.module';
-export { YUPModule } from './src/yup/yup.module';
+// export { YUPModule } from './src/yup/backup/yup.module';
+export { YupModule, Yup, YUP_DATA, YupRef } from './src/yup';
 export { YUPService } from './src/yup/yup.service';
 export { AudioModule } from './src/yudio/module';
 export { AudioService } from './src/yudio/service';