+++ src/yup/index.ts
@@ -1,0 +1,3 @@
+export { YupModule } from './module';
+export { DialogService as Yup } from './service';
+export { YupRef, YUP_DATA } from './popup.ref';