+++ src/yup/templates/index.ts
@@ -1,0 +1,4 @@
+export { DialogComponent } from './dialog';
+export { AlertComponent } from './alert';
+export { LoadComponent } from './load';
+export { ToastComponent } from './toast';