+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.5.0 (2017-09-14)
+### Fix
+*nothing*
+### Refactoring
+* refact default popups(alert, dialog, toast, loader)
+* seems a little more like material2 dialog creating
+### Features
+* now yup can use open() method to create a custom angular component and add it at the end of the body tag
+### Breaking Changes
+* now yup module no more need to add a \<yup\> tag, the popup dom will be added to the end of body tag
+
 ## 1.4.2 (2017-09-13)
 ### Fix
 * fix peerDependencies, remove package for Material2